- fix product id and shop id data type

+++ libraries/cm_push_notifications/src/main/java/com/tokopedia/notifications/model/AddToCart.kt
@@ -7,37 +7,37 @@ import com.google.gson.annotations.SerializedName
 import com.tokopedia.notifications.common.CMConstant
 
 data class AddToCart(
-        @Expose @SerializedName(CMConstant.PayloadKeys.PRODUCT_ID) var productId: Int? = 0,
+        @Expose @SerializedName(CMConstant.PayloadKeys.PRODUCT_ID) var productId: Long? = 0,
         @Expose @SerializedName(CMConstant.PayloadKeys.PRODUCT_NAME) var productName: String? = "",
         @Expose @SerializedName(CMConstant.PayloadKeys.PRODUCT_BRAND) var productBrand: String? = "",
         @Expose @SerializedName(CMConstant.PayloadKeys.PRODUCT_PRICE) var productPrice: Float? = 0f,
         @Expose @SerializedName(CMConstant.PayloadKeys.PRODUCT_VARIANT) var productVariant: String? = "",
         @Expose @SerializedName(CMConstant.PayloadKeys.PRODUCT_QUANTITY) var productQuantity: String? = "",
-        @Expose @SerializedName(CMConstant.PayloadKeys.ATC_SHOP_ID) var shopId: Int? = 0,
+        @Expose @SerializedName(CMConstant.PayloadKeys.ATC_SHOP_ID) var shopId: Long? = 0,
         @Expose @SerializedName(CMConstant.PayloadKeys.SHOP_NAME) var shopName: String? = "",
         @Expose @SerializedName(CMConstant.PayloadKeys.SHOP_TYPE) var shopType: String? = ""
 ) : Parcelable {
 
     constructor(parcel: Parcel) : this(
-            parcel.readInt(),
+            parcel.readLong(),
             parcel.readString(),
             parcel.readString(),
             parcel.readFloat(),
             parcel.readString(),
             parcel.readString(),
-            parcel.readInt(),
+            parcel.readLong(),
             parcel.readString(),
             parcel.readString()
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(productId?: 0)
+        parcel.writeLong(productId?: 0)
         parcel.writeString(productName)
         parcel.writeString(productBrand)
         parcel.writeFloat(productPrice?: 0f)
         parcel.writeString(productVariant)
         parcel.writeString(productQuantity)
-        parcel.writeInt(shopId?: 0)
+        parcel.writeLong(shopId?: 0)
         parcel.writeString(shopName)
         parcel.writeString(shopType)
     }
